Split fork additions to stay within file length budgets
Move SessionRestorableAgentSnapshot (and its private script store) out of
RestorableAgentSession.swift, move the new fork tests into
AgentConversationForkCommandTests.swift, and fold the .custom palette
availability case into the existing supported list.
